genai-feat-private_assistant/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.8.2
pyyaml==6.0.1
gradio==4.44.1
python-dotenv==1.0.1
# optional privacy
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
